+++ authentication_ui_bootstrap5/src/main/java/org/octri/authentication/ui/bootstrap5/DocumentationDummy.java
@@ -1,0 +1,8 @@
+package org.octri.authentication.ui.bootstrap5;
+
+/**
+ * Dummy class to ensure that a Javadoc jar is generated for the Bootstrap UI package.
+ */
+public class DocumentationDummy {
+
+}